refactor: fix & add a bunch of stuff b/c i'm crazy

+++ fish/functions/finder/hidefiles.fish
@@ -1,3 +1,8 @@
 function hidefiles -d 'Hide the normally hidden dotfiles in Finder'
+    if test (uname -s) != Darwin
+        echo >&2 "Cannot execute command. MacOS not detected."
+        return 1
+    end
+
     defaults write com.apple.finder AppleShowAllFiles -bool false && killall Finder
 end
