+++ fish/functions/networking/flushdns.fish
@@ -1,3 +1,8 @@
 function flushdns -d 'Flush DNS cache'
+    if test (uname -s) != Darwin
+        echo >&2 "Cannot execute command. MacOS not detected."
+        return 1
+    end
+
     dscacheutil -flushcache && sudo killall -HUP mDNSResponder
 end
